Reference

Main functions

RedClust.fitprior — Function

fitprior(data, algo, dist = false, Kmin = 1, Kmax = Int(floor(size(data)[end] / 2))

Compute the prior hyperparameters from the data.

Arguments

data::Union{Vector{Vector{Float64}}, Matrix{Float64}}: can either be a vector of (possibly multi-dimensional) observations, or a matrix with each column an observation, or a square matrix of pairwise dissimilarities.

algo::String: must be one of "k-means" or "k-medoids".

dist::bool = false: if true, data will be assumed to be a pairwise dissimilarity matrix.

Kmin::Integer: minimum number of clusters.

Kmax::Integer = Int(floor(size(data)[end] / 2)): maximum number of clusters. If left unspecified, it is set to half the number of observations.

Returns

A named tuple containing the following

params::PriorHyperparamsList: the fitted hyperparameters.

K::Integer: the number of clusters in the notional clustering.

notionalclustering::ClustLabelVector: the cluster labels in the notional clustering.

See also

PriorHyperparamsList

RedClust.runsampler — Function

runsampler(data[, options, params])

Runs the MCMC sampler on the data.

Arguments

data::MCMCData: contains the distance matrix.

options = MCMCOptionsList(): contains the number of iterations, burnin, etc.

params = PriorHyperparamsList(): contains the prior hyperparameters for the model.

Returns

A struct of type MCMCResult containing the MCMC samples, convergence diagnostics, and summary statistics.

See also

MCMCData, MCMCOptionsList, fitprior, PriorHyperparamsList.

Summary and clustering evaluation

RedClust.evaluateclustering — Method

evaluateclustering(clusts::ClustLabelVector, truth::ClustLabelVector)

Returns a tuple of values quantifying the accuracy of the clustering assignments in clusts with respect to the ground truth clustering assignments in truth.

Return values

bloss: Normalised Binder loss (= 1 - rand index). Lower is better.

ari: Adjusted Rand index. Higher is better.

vidist: Variation of Information distances. Lower is better.

vmeas: V-Measure. Higher is better.

nmi: Normalised Mutual Information. Higher is better.

RedClust.summarise — Function

summarise(result::MCMCResult, truth::ClustLabelVector = [], printoutput = true)

Returns a string that summarises the MCMC output. If truth is a vector of cluster labels, the output also contains clustering metrics to quantify the accuracy of the point-estimate from the MCMC with respect to truth. If printoutput = true, the output will be printed to console before being returned.

Convenience functions

RedClust.adjacencymatrix — Method

adjacencymatrix(clusts::ClustLabelVector)

Returns the nxn adjacency matrix corresponding to the given cluster label vector clusts, where n = length(clusts).

RedClust.generatemixture — Method

generatemixture(N, K; α = K, dim = K, radius = 1, σ = 0.1)

Generates a multivariate Normal mixture, with kernel weights generated from a Dirichlet prior. The kernels are centred at the vertices of a dim-dimensional simplex with edge length radius.

Arguments

N::Integer: number of observations to generate.

K::Integer: number of mixture kernels.

α::Float64 = K: parameter for the Dirichlet prior.

dim::Integer = K: dimension of the observations.

radius::Float64 = 1: radius of the simplex whose vertices are the kernel means.

σ::Float64 = 0.1: variance of each kernel.

Returns

Named tuple containing the following fields-

pnts::Vector{Vector{Float64}}: a vector of N observations.

distM::Matrix{Float64}: an NxN matrix of pairwise Euclidean distances between the observations.

clusts::ClustLabelVector: vector of N cluster assignments.

probs::Float64: vector of K cluster weights generated from the Dirichlet prior, used to generate the observations.

oracle_coclustering::Matrix{Float64}: NxN matrix of co-clustering probabilities, calculated assuming full knowledge of the cluster centres and cluster weights.

RedClust.makematrix — Method

Convert a vector of vectors into a matrix, where each vector becomes a column in the matrix.

RedClust.sortlabels — Method

sortlabels(x::ClustLabelVector)

Returns a cluster label vector y such that x and y have the same adjacency structure and labels in y occur in sorted ascending order.

Types

RedClust.ClustLabelVector — Type

Alias for Vector{Int}

RedClust.MCMCData — Type

Contains the pairwise dissimilarities. Construct by calling MCMCData(D), where D is a square matrix of pairwise dissimilarities.

RedClust.MCMCOptionsList — Type

List of options for running the MCMC.

Constructor: MCMCOptionsList(; [numiters, burnin, thin, numGibbs, numMH, usesalso])

Constructor arguments

numiters::Integer = 5000: number of iterations to run.

burnin::Integer = floor(0.2 * numiters): number of iterations to discard as burn-in.

thin::Integer = 1: will keep every thin samples.

numGibbs:Integer = 1: number of intermediate Gibbs scans in the split-merge step.

numMH:Integer = 1: number of split-merge steps per MCMC iteration.

usesalso::Bool = true: whether to use the SALSO algorithm to compute a point estimate.

RedClust.MCMCResult — Type

Struct containing MCMC samples.

Fields

clusts::Vector{ClustLabelVector}: contains the clustering allocations. clusts[i] is a vector containing the clustering allocation from the ith sample.

pntestimate::ClustLabelVector}: contains the point-estimate clustering allocation.

posterior_coclustering::Matrix{Float64}: the posterior coclustering matrix.

K::Vector{Int}: posterior samples of $K$, i.e., the number of clusters. K[i] is the number of clusters in clusts[i].

r::Vector{Float64}: posterior samples of the parameter $r$.

p::Vector{Float64}: posterior samples of the parameter $p$.

K_mean, r_mean, p_mean: posterior mean of K, r, and p respectively.

K_variance, r_variance, p_variance: posterior variance of K, r, and p respectively.

K_acf::Vector{Float64}, r_acf::Vector{Float64}, p_acf::Vector{Float64}: autocorrelation function for K, r, and p respectively.

K_iac, r_iac, and p_iac: integrated autocorrelation coefficient for K, r, and p respectively.

K_ess::Float64, r_ess::Float64, and p_ess::Float64: effective sample size for K, r, and p respectively.

loglik::Vector{Float64}: log-likelihood for each sample.

splitmerge_splits: Boolean vector indicating the iterations when a split proposal was used in the split-merge step. Has length numMH * numiters (see MCMCOptionsList).

splitmerge_acceptance_rate: acceptance rate of the split-merge proposals.

r_acceptances: Boolean vector indicating the iterations (including burnin and the thinned out iterations) where the Metropolis-Hastings proposal for r was accepted.

r_acceptance_rate:: Metropolis-Hastings acceptance rate for r.

runtime: total runtime for all iterations.

mean_iter_time: average time taken for each iteration.

options::MCMCOptionsList: options passed to the sampler.

params::PriorHyperparamsList: prior hyperparameters used by the sampler.

RedClust.PriorHyperparamsList — Type

Contains the prior hyperparameters for the model. Call the constructor with the line

PriorHyperparamsList(; [kwargs])

Constructor arguments

δ1::Float64 = 1: the parameter $\delta_1$.

α::Float64 = 1: the shape parameter $\alpha$ in the prior for each $\lambda_k$.

β::Float64 = 1: the rate parameter $\beta$ in the prior for each $\lambda_k$.

δ2::Float64 = 1: the parameter $\delta_2$.

ζ::Float64 = 1: the shape parameter $\zeta$ in the prior for each $\theta_{kt}$.

γ::Float64 = 1: the rate parameter $\gamma$ in the prior for each $\theta_{kt}$.

η::Float64 = 1: the shape parameter $\eta$ in the prior for $r$.

σ::Float64 = 1: the rate parameter $\sigma$ in the prior for $r$.

u::Float64 = 1: the parameter $u$ in the prior for $p$.

v::Float64 = 1: the parameter $v$ in the prior for $p$.

proposalsd_r::Float64 = 1: standard deviation of the truncated Normal proposal when sampling r via a Metropolis-Hastings step.

K_initial::Int = 1: initial number of clusters to start the MCMC.

repulsion::Bool = true: whether to use the repulsive component of the likelihood.

maxK::Int = 0: maxmimum number of clusters to allow. If set to 0 then we can get up to n clusters, where n is the number of observations.

See also

fitprior